+++ encoding/thrift/options.go
@@ -23,6 +23,7 @@ package thrift
 import "go.uber.org/thriftrw/protocol"
 
 type clientConfig struct {
+	ServiceName string
 	Protocol    protocol.Protocol
 	Enveloping  bool
 	Multiplexed bool
@@ -34,8 +35,9 @@ type ClientOption interface {
 }
 
 type registerConfig struct {
-	Protocol   protocol.Protocol
-	Enveloping bool
+	ServiceName string
+	Protocol    protocol.Protocol
+	Enveloping  bool
 }
 
 // RegisterOption customizes the behavior of a Thrift handler during
@@ -91,6 +93,41 @@ type multiplexedOption struct{}
 
 func (multiplexedOption) applyClientOption(c *clientConfig) {
 	c.Multiplexed = true
+}
+
+type namedOption struct{ ServiceName string }
+
+func (n namedOption) applyClientOption(c *clientConfig) {
+	if c.ServiceName == "" {
+		c.ServiceName = n.ServiceName
+	}
+}
+
+func (n namedOption) applyRegisterOption(c *registerConfig) {
+	if c.ServiceName == "" {
+		c.ServiceName = n.ServiceName
+	}
+}
+
+// Named is an option that specifies the name of a thrift.Client.
+// This option should be used if a thrift service extends another
+// thrift service. Note that the first Named Option will trump all
+// other Named options. This ensures that the inherited procedures
+// are appropriately labelled with the correct inheriting service's
+// name.
+//
+// Specify this option when constructing the Thrift client.
+//
+//  client := myserviceclient.New(clientConfig, thrift.Named("foo"))
+//
+// It may be specified on the server side when the handler is registered.
+//
+// 	dispatcher.Register(myserviceserver.New(handler, thrift.Named("foo")))
+//
+// If not specified, the client or server's inherited procedures will
+// be labelled with the service name from which they are inherited.
+func Named(n string) Option {
+	return namedOption{ServiceName: n}
 }
 
 type protocolOption struct{ Protocol protocol.Protocol }