+++ encoding/thrift/thriftrw-plugin-yarpc/internal/tests/atomic/readonlystoreserver/server.go
@@ -47,7 +47,17 @@ func New(impl Interface, opts ...thrift.RegisterOption) []transport.Procedure {
 	}
 
 	procedures := make([]transport.Procedure, 0, 1)
-	procedures = append(procedures, baseserviceserver.New(impl, opts...)...)
+
+	procedures = append(
+		procedures,
+		baseserviceserver.New(
+			impl,
+			append(
+				opts,
+				thrift.Named("ReadOnlyStore"),
+			)...,
+		)...,
+	)
 	procedures = append(procedures, thrift.BuildProcedures(service, opts...)...)
 	return procedures
 }