+++ encoding/thrift/thriftrw-plugin-yarpc/internal/tests/atomic/storeclient/client.go
@@ -47,7 +47,14 @@ func New(c transport.ClientConfig, opts ...thrift.ClientOption) Interface {
 			Service:      "Store",
 			ClientConfig: c,
 		}, opts...),
-		Interface: readonlystoreclient.New(c, opts...),
+
+		Interface: readonlystoreclient.New(
+			c,
+			append(
+				opts,
+				thrift.Named("Store"),
+			)...,
+		),
 	}
 }