+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.34.0] - 2018-10-03
+### Added
+- Adds `thrift.Named` option for appropriately labelling procedures inherited
+  from other thrift services.
+- The HTTP protocol now marks peers as unavailable immediately when the remote
+side closes the connection.
+
+### Fixed
+- Calling extended Thrift service procedures previously called the base service's
+  procedures.
+
 ## [1.33.0] - 2018-09-26
 ### Added
 - x/yarpctest: Add a retry option to HTTP/TChannel/GRPCRequest.
@@ -1011,6 +1022,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[1.34.0]: https://github.com/yarpc/yarpc-go/compare/v1.33.0...v1.34.0
 [1.33.0]: https://github.com/yarpc/yarpc-go/compare/v1.32.4...v1.33.0
 [1.32.4]: https://github.com/yarpc/yarpc-go/compare/v1.32.3...v1.32.4
 [1.32.3]: https://github.com/yarpc/yarpc-go/compare/v1.32.2...v1.32.3